Extract StatsRow component as a file
