+++ app/src/main/java/com/jp_funda/boxiful/views/record/component/WeeklyStatsSection.kt
@@ -1,0 +1,12 @@
+package com.jp_funda.boxiful.views.record.component
+
+import androidx.compose.runtime.Composable
+import androidx.hilt.navigation.compose.hiltViewModel
+import com.jp_funda.boxiful.views.record.RecordViewModel
+
+@Composable
+fun WeeklyStatsSection() {
+    val viewModel = hiltViewModel<RecordViewModel>()
+
+    
+}
